# scal benchmark: eliminate y, move init/timing out of loop