@bartoldeman 2022-11-29
Two notes/questions:
1. Many other benchmarks have the init in the loop, but not gemm. I couldn't figure out why? Should they be adjusted too?
2. For some reason I don't understand `kernel/x86_64/dscal_microk_haswell-2.c` uses xmm registers, where ymm registers are twice as fast on Zen2 (also Broadwell). Even stranger, the sscal kernel uses the old code in `scal_sse.S`, unlike [dcz]. Do you know why? I can supply PRs for those too.