@martin-frbg 2022-11-29
1. Could be simply down to when the benchmarks were added - most appear to have been written by the same contributor in 2014 (and some of them were later used as templates for others), and he wrote the GEMM one years later, when he may have been much more aware of cache effects.
2. The Haswell kernel precedes the introduction of Ryzen2 by three years, and probably nobody was aware of the register speed issue. I have no idea if the absence of microkernels for SSCAL is/was due to limited benefit expected from a newer-than-SSE2 solution, or if it was never seen as a bottleneck in real-world code. Unfortunately the author of the files in question stopped contributing very suddenly and for unknown (possibly even medical) reasons in early 2017.