@brada4 2022-12-01
Level 1 BLAS is bound to memory speed - you get like few flops per memory access. Better characteristic would be GB/s  towards RAM in L1 and in many cases for L2,but for L3 is more like N^(3/2) per memory access. 

Not here, but for other benchmarks memset()-ing destination allocation is needed to cancel out memory overcommitment effects.

Best place to start would be the benchmarks which are matching python/julia/R scripts in subdirectory.