@brada4 2022-12-01
More like X cachelines per memory cycle, but ok, we are  on similar page.